removed FIXME
I left a FIXME saying that a substitution to convert a Linux path
to a Windows path for TASM32 would not work if the path was absolute,
since there is no drive letter - but this was only an assumption.
I just tested it, and it actually does work, presumably assuming
the current drive, which, by default, is Z:. As a result, I have
removed the FIXME to avoid confusion.

Signed-off-by: RogueAI42 <declanhoare@exemail.com.au>